Optimize loading speed by removing addressSingleTokenBalancesQuery
